switch to shell-based config
`tmux` config parsing is not very robust and seems export an unstable
API. This change switches to a shell-based architecture while retaining
the same commands.

This change was made due to issues with tmux configuration parsing. The
configuration grammar has changed drastically from 3.1 to 3.2. Composing
complex commands has become more difficult and, while the base set of
commands are still supported, it's become harder to support more complex
operations. The goal of this project is ultimately to replicate as much
`dwm` functionality as possible and a shell-based approach is, for now,
the only way to accomplish that goal.

Some notes:

- `dwm.tmux` bin handles all commands
- `dwm.tmux` lib sets up a `command-alias` and default keybindings for
  each command. `command-alias` does not support alias chaining, so
  something like this is not possible:
  ```
  set -g command-alias[10] foo='run-shell ls'
  set -g command-alias[11] bar='foo'
  ```
- each `command-alias` uses `run-shell 'dwm.tmux $command'` to execute
  the appropriate command
- `run-shell` adds about 20ms of additional overhead to each command
- `shc` was tested to see if it could reduce the overhead but no change
  was shown. This may be evaluated again later