add mfact support
This change adds support for increasing and decreasing the size of the
main pane, or `mfact`. Two new commands are introduced, `incmfact` and
`decmfact`, which respecitively increase and decrease the `mfact` value.
The value range is 5..95 and increments by 5. Pane commands respect the
value.

The default value can be set in a configuration file using `setenv`.
Changes made by the `*mfact` commands act on the session value, not the
global value.