Simplify "set -- ..." and fix running on Bash
Bash breaks with POSIX in that it does not interpret "\n" in a simple
"echo". For Bash, it would have to be "echo -e ...".

But "echo" is not needed here anyway. Just separate the variables by
spaces and let the shell's word splitting routines do the work (which is
the same as before, actually, just without an extra step).